+++ covalent_awslambda_plugin/awslambda.py
@@ -19,16 +19,12 @@
 # Relief from the License may be granted by purchasing a commercial license.
 
 import asyncio
+import json
 import os
-import pathlib
-import shutil
-import subprocess
-import sys
 import tempfile
 import time
 from contextlib import contextmanager
 from typing import Callable, Dict, List, Tuple
-from zipfile import ZipFile
 
 import boto3
 import botocore.exceptions
@@ -38,14 +34,13 @@ from covalent._shared_files import logger
 from covalent._shared_files.config import get_config
 from covalent_aws_plugins import AWSExecutor
 
-from .scripts import PYTHON_EXEC_SCRIPT
-
 app_log = logger.app_log
 log_stack_info = logger.log_stack_info
 
 executor_plugin_name = "AWSLambdaExecutor"
 
 _EXECUTOR_PLUGIN_DEFAULTS = {
+    "function_name": "covalent-awslambda-executor",
     "credentials_file": os.environ.get("AWS_SHARED_CREDENTIALS_FILE")
     or os.path.join(os.environ.get("HOME"), ".aws/credentials"),
     "profile": os.environ.get("AWS_PROFILE") or "default",
@@ -58,87 +53,15 @@ _EXECUTOR_PLUGIN_DEFAULTS = {
     "cleanup": False,
 }
 
-LAMBDA_FUNCTION_NAME = "lambda-{dispatch_id}-{node_id}"
 FUNC_FILENAME = "func-{dispatch_id}-{node_id}.pkl"
 RESULT_FILENAME = "result-{dispatch_id}-{node_id}.pkl"
-LAMBDA_DEPLOYMENT_ARCHIVE_NAME = "archive-{dispatch_id}-{node_id}.zip"
-LAMBDA_FUNCTION_SCRIPT_NAME = "lambda_function.py"
-
-
-class DeploymentPackageBuilder:
-    """AWS Lambda deployment package (zip archive) builder
-
-    Args:
-        directory: Path to local filesystem on where to store the archive
-        archive_name: Name of the deployment archive
-        s3_bucket_name: Name of the AWS S3 bucket to be used to cache the deployment package
-    """
-
-    def __init__(self, directory: str, archive_name: str, s3_bucket_name: str):
-        self.directory = directory
-        self.archive_name = archive_name
-        self.s3_bucket_name = s3_bucket_name
-        self.target_dir = os.path.join(self.directory, "targets")
-        self.deployment_archive = os.path.join(self.directory, self.archive_name)
-
-    async def install(self, pkg_name: str):
-        """Install the necessary Python packages into the specified target directory
-
-        Args:
-            pkg_name: Name of the Python package to be installed
-
-        Returns:
-            None
-        """
-        cmd = " ".join(
-            [
-                sys.executable,
-                "-m",
-                "pip",
-                "install",
-                "--target",
-                self.target_dir,
-                "--upgrade",
-                pkg_name,
-            ]
-        )
-        proc, stdout, stderr = await AWSLambdaExecutor.run_async_subprocess(cmd)
-        if proc.returncode != 0:
-            app_log.error(stderr)
-            raise RuntimeError(f"Unable to install package {pkg_name}")
-
-    def write_deployment_archive(self):
-        # Create zip archive with dependencies
-        with ZipFile(self.deployment_archive, mode="w") as archive:
-            for file_path in pathlib.Path(self.target_dir).rglob("*"):
-                archive.write(
-                    file_path, arcname=file_path.relative_to(pathlib.Path(self.target_dir))
-                )
-
-    async def __aenter__(self):
-        """Create the zip archive"""
-        if os.path.exists(self.target_dir):
-            shutil.rmtree(self.target_dir)
-            os.mkdir(self.target_dir)
-
-        # Install the required python dependencies
-        await self.install("boto3")
-        await self.install("covalent==0.177.0.post1.dev0")
-
-        loop = asyncio.get_running_loop()
-        await loop.run_in_executor(None, self.write_deployment_archive)
-
-        return self.deployment_archive
-
-    async def __aexit__(self, exc_type, exc_value, exc_tb):
-        """None"""
-        pass
 
 
 class AWSLambdaExecutor(AWSExecutor):
     """AWS Lambda executor plugin
 
     Args:
+        function_name: Name of an existing lambda function to use during execution (default: `covalent-awsambda-executor`)
         credentials_file: Path to AWS credentials file (default: `~/.aws/credentials`)
         profile: AWS profile (default: `default`)
         region: AWS region (default: `us-east-1`)
@@ -151,6 +74,7 @@ class AWSLambdaExecutor(AWSExecutor):
 
     def __init__(
         self,
+        function_name: str,
         s3_bucket_name: str,
         execution_role: str,
         credentials_file: str = None,
@@ -175,12 +99,15 @@ class AWSLambdaExecutor(AWSExecutor):
         super().__init__(**required_attrs)
 
         # Lambda executor parameters
+        self.function_name = (
+            function_name
+            or get_config("executors.awslambda.function_name")
+            or "covalent-awslambda-executor"
+        )
         self.poll_freq = poll_freq or get_config("executors.awslambda.poll_freq")
         self.timeout = timeout or get_config("executors.awslambda.timeout")
         self.memory_size = memory_size or get_config("executors.awslambda.memory_size")
         self.cleanup = cleanup
-
-        self._cwd = tempfile.mkdtemp()
         self._key_exists = False
 
     @contextmanager
@@ -223,77 +150,9 @@ class AWSLambdaExecutor(AWSExecutor):
         fut = loop.run_in_executor(None, self._upload_task_sync, workdir, func_filename)
         await fut
 
-    async def _is_lambda_active(self, function_name: str) -> bool:
-        """Check if the lambda function is active of not
-
-        Args:
-            function_name: Name of the lambda function
-
-        Returns:
-            bool: Boolean value either True/False
-        """
-        # Check if lambda is active
-        is_active = False
-        with self.get_session() as session:
-            lambda_client = session.client("lambda")
-            while not is_active:
-                lambda_state = lambda_client.get_function(FunctionName=function_name)
-                app_log.debug(
-                    f"Lambda function {function_name} state: {lambda_state['Configuration']['State']}"
-                )
-                if lambda_state["Configuration"]["State"] == "Active":
-                    is_active = True
-                else:
-                    await asyncio.sleep(0.5)
-                    continue
-        return is_active
-
-    async def _create_lambda(self, function_name: str, deployment_archive_name: str) -> Dict:
-        """Create the AWS Lambda function
-
-        Args:
-            function_name: AWS Lambda function name
-            deployment_archive_name: Lambda deployment zip archive name
-
-        Returns:
-            response: AWS boto3 client create_lambda response
-        """
-        with self.get_session() as session:
-            iam_client = session.client("iam")
-            role_arn = None
-            try:
-                response = iam_client.get_role(RoleName=f"{self.execution_role}")
-                role_arn = response["Role"]["Arn"]
-            except botocore.exceptions.ClientError as ce:
-                app_log.exception(ce)
-                raise
-
-        lambda_create_kwargs = {
-            "FunctionName": f"{function_name}",
-            "Runtime": "python3.8",
-            "Role": f"{role_arn}",
-            "Handler": "lambda_function.lambda_handler",
-            "Code": {"S3Bucket": f"{self.s3_bucket_name}", "S3Key": deployment_archive_name},
-            "PackageType": "Zip",
-            "Publish": True,
-            "Timeout": self.timeout,
-            "MemorySize": self.memory_size,
-        }
-
-        with self.get_session() as session:
-            lambda_client = session.client("lambda")
-            try:
-                response = lambda_client.create_function(**lambda_create_kwargs)
-                app_log.debug(f"Lambda function: {function_name} created: {response}")
-            except botocore.exceptions.ClientError as ce:
-                app_log.exception(ce)
-                raise
-
-        # Check if lambda is active
-        is_active = await self._is_lambda_active(function_name)
-        return "Active" if is_active else None
-
-    def submit_task_sync(self, function_name: str) -> Dict:
+    def submit_task_sync(
+        self, function_name: str, func_filename: str, result_filename: str
+    ) -> Dict:
         """The actual (blocking) submit_task function"""
 
         app_log.debug(f"Invoking AWS Lambda function {function_name}")
@@ -301,12 +160,23 @@ class AWSLambdaExecutor(AWSExecutor):
         with self.get_session() as session:
             client = session.client("lambda")
             try:
-                return client.invoke(FunctionName=function_name)
+                return client.invoke(
+                    FunctionName=function_name,
+                    Payload=json.dumps(
+                        {
+                            "S3_BUCKET_NAME": self.s3_bucket_name,
+                            "COVALENT_TASK_FUNC_FILENAME": func_filename,
+                            "RESULT_FILENAME": result_filename,
+                        }
+                    ),
+                )
             except botocore.exceptions.ClientError as ce:
                 app_log.exception(ce)
                 raise
 
-    async def submit_task(self, function_name: str) -> Dict:
+    async def submit_task(
+        self, function_name: str, func_filename: str, result_filename: str
+    ) -> Dict:
         """
         Submit the task by invoking the AWS Lambda function
 
@@ -317,7 +187,9 @@ class AWSLambdaExecutor(AWSExecutor):
             response: AWS boto3 client invoke lambda response
         """
         loop = asyncio.get_running_loop()
-        fut = loop.run_in_executor(None, self.submit_task_sync, function_name)
+        fut = loop.run_in_executor(
+            None, self.submit_task_sync, function_name, func_filename, result_filename
+        )
         return await fut
 
     def get_status_sync(self, object_key: str):
@@ -396,87 +268,10 @@ class AWSLambdaExecutor(AWSExecutor):
 
         return result_object
 
-    def _upload_deployment_archive_sync(self, deployment_archive, deployment_archive_name):
-        with self.get_session() as session:
-            client = session.client("s3")
-            try:
-                client.upload_file(
-                    deployment_archive, self.s3_bucket_name, deployment_archive_name
-                )
-            except botocore.exceptions.ClientError as ce:
-                app_log.exception(ce)
-                raise
-
     async def query_result(self, workdir: str, result_filename: str):
         loop = asyncio.get_running_loop()
         fut = loop.run_in_executor(None, self.query_result_sync, workdir, result_filename)
         return await fut
-
-    async def setup(self, task_metadata: Dict):
-        """AWS Lambda specific setup tasks
-
-        Args:
-            task_metadata: Dictionary containing the task dispatch_id and node_id
-
-        Returns:
-            None
-        """
-        dispatch_id = task_metadata["dispatch_id"]
-        node_id = task_metadata["node_id"]
-        workdir = os.path.join(self._cwd, dispatch_id)
-        deployment_archive_name = LAMBDA_DEPLOYMENT_ARCHIVE_NAME.format(
-            dispatch_id=dispatch_id, node_id=node_id
-        )
-        lambda_function_name = LAMBDA_FUNCTION_NAME.format(
-            dispatch_id=dispatch_id, node_id=node_id
-        )
-
-        app_log.debug(f"Starting setup for task - {dispatch_id}-{node_id} ... ")
-
-        if not os.path.exists(workdir):
-            os.mkdir(workdir)
-
-        app_log.debug(
-            f"Creating the Lambda deployment archive at {os.path.join(workdir, deployment_archive_name)} ..."
-        )
-        async with DeploymentPackageBuilder(
-            workdir, deployment_archive_name, self.s3_bucket_name
-        ) as deployment_archive:
-            exec_script = PYTHON_EXEC_SCRIPT.format(
-                func_filename=FUNC_FILENAME.format(dispatch_id=dispatch_id, node_id=node_id),
-                result_filename=RESULT_FILENAME.format(dispatch_id=dispatch_id, node_id=node_id),
-                s3_bucket_name=self.s3_bucket_name,
-            )
-            with open(os.path.join(workdir, LAMBDA_FUNCTION_SCRIPT_NAME), "w") as f:
-                f.write(exec_script)
-            # Add script to the deployment archive
-            with ZipFile(deployment_archive, mode="a") as archive:
-                archive.write(
-                    os.path.join(workdir, LAMBDA_FUNCTION_SCRIPT_NAME),
-                    arcname=LAMBDA_FUNCTION_SCRIPT_NAME,
-                )
-
-        app_log.debug(
-            f"Lambda deployment archive: {os.path.join(workdir, deployment_archive_name)} created. Uploading to S3 ..."
-        )
-
-        # Upload archive to s3 bucket
-        loop = asyncio.get_running_loop()
-        await loop.run_in_executor(
-            None,
-            self._upload_deployment_archive_sync,
-            deployment_archive,
-            deployment_archive_name,
-        )
-
-        app_log.debug(f"Lambda deployment archive: {deployment_archive_name} uploaded to S3 ... ")
-
-        # Create the lambda function
-        app_log.debug(f"Creating AWS Lambda function {lambda_function_name} ...")
-        state = await self._create_lambda(lambda_function_name, deployment_archive_name)
-        app_log.debug(f"Lambda function: {lambda_function_name} created in state: {state}")
-
-        app_log.debug(f"Finished setup for task - {dispatch_id}-{node_id} ... ")
 
     async def run(self, function: Callable, args: List, kwargs: Dict, task_metadata: Dict):
         """Run the executor
@@ -492,14 +287,10 @@ class AWSLambdaExecutor(AWSExecutor):
         """
         dispatch_id = task_metadata["dispatch_id"]
         node_id = task_metadata["node_id"]
-        workdir = os.path.join(self._cwd, dispatch_id)
+        workdir = self.cache_dir
 
         func_filename = FUNC_FILENAME.format(dispatch_id=dispatch_id, node_id=node_id)
         result_filename = RESULT_FILENAME.format(dispatch_id=dispatch_id, node_id=node_id)
-        lambda_function_name = LAMBDA_FUNCTION_NAME.format(
-            dispatch_id=dispatch_id, node_id=node_id
-        )
-
         app_log.debug(f"In run for task - {dispatch_id} - {node_id} ... ")
 
         app_log.debug("Pickling function, args and kwargs..")
@@ -510,7 +301,9 @@ class AWSLambdaExecutor(AWSExecutor):
         await self._upload_task(workdir, func_filename)
 
         # Invoke the created lambda
-        lambda_invocation_response = await self.submit_task(lambda_function_name)
+        lambda_invocation_response = await self.submit_task(
+            self.function_name, func_filename, result_filename
+        )
         app_log.debug(f"Lambda function response: {lambda_invocation_response}")
         if "FunctionError" in lambda_invocation_response:
             error = lambda_invocation_response["Payload"].read().decode("utf-8")
@@ -533,75 +326,6 @@ class AWSLambdaExecutor(AWSExecutor):
         Cancel execution
         """
         raise NotImplementedError("Cancellation is currently not supported")
-
-    async def teardown(self, task_metadata: Dict):
-        """Cleanup temporary files and the Lambda function
-
-        Args:
-            task_metadata: Dictionary containing the task dispatch_id and node_id
-
-        Returns:
-            None
-        """
-        dispatch_id = task_metadata["dispatch_id"]
-        node_id = task_metadata["node_id"]
-        lambda_function_name = LAMBDA_FUNCTION_NAME.format(
-            dispatch_id=dispatch_id, node_id=node_id
-        )
-        workdir = os.path.join(self._cwd, dispatch_id)
-
-        app_log.debug(f"In teardown for task - {dispatch_id} - {node_id}")
-
-        if self.cleanup:
-            with self.get_session() as session:
-
-                app_log.debug(f"Cleaning up resources created in S3 bucket {self.s3_bucket_name}")
-                s3_resource = session.resource("s3")
-
-                try:
-                    s3_resource.Object(
-                        self.s3_bucket_name,
-                        FUNC_FILENAME.format(dispatch_id=dispatch_id, node_id=node_id),
-                    ).delete()
-
-                    s3_resource.Object(
-                        self.s3_bucket_name,
-                        RESULT_FILENAME.format(dispatch_id=dispatch_id, node_id=node_id),
-                    ).delete()
-
-                    s3_resource.Object(
-                        self.s3_bucket_name,
-                        LAMBDA_DEPLOYMENT_ARCHIVE_NAME.format(
-                            dispatch_id=dispatch_id, node_id=node_id
-                        ),
-                    ).delete()
-
-                    app_log.debug(f"All objects from bucket {self.s3_bucket_name} deleted")
-
-                except botocore.exceptions.ClientError as ce:
-                    app_log.exception(ce)
-                    raise
-
-                app_log.debug("Cleanup of resources in S3 bucket finished")
-
-                app_log.debug(f"Deleting lambda function {lambda_function_name}")
-                lambda_client = session.client("lambda")
-
-                try:
-                    response = lambda_client.delete_function(FunctionName=lambda_function_name)
-                    app_log.debug(f"Lambda cleanup response: {response}")
-                except botocore.exceptions.ClientError as ce:
-                    app_log.exception(ce)
-                    raise
-
-                app_log.debug(f"Lambda function {lambda_function_name} deleted")
-
-                app_log.debug(f"Cleaning up working directory {workdir}")
-                if os.path.exists(workdir):
-                    shutil.rmtree(workdir)
-                app_log.debug(f"Working directory {workdir} deleted")
-
-        app_log.debug(f"Finished teardown for task - {dispatch_id} - {node_id}")
 
     # copied from RemoteExecutor
     @staticmethod